separating What We Offer into separate component
