+++ Haven/src/components/WhatWeOfferModal/what-we-offer-modal.css
@@ -9,7 +9,7 @@
   align-items: center;
   justify-content: center;
   animation-name: animateModal;
-  animation-duration: 0.6s;
+  animation-duration: 0.4s;
 }
 
 .modal-box {
