+++ Haven/src/pages/Bookings/bookings.css
@@ -9,17 +9,6 @@
   padding-bottom: 20px;
 }
 
-/* .top-header-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 0px;
-  margin: 0px;
-  padding: 0px;
-  width: 100%;
-} */
-
 .bookings-container h3 {
   width: 900px;
 }
@@ -74,7 +63,7 @@
   align-items: center;
 }
 
-.what-we-offer {
+/* .what-we-offer {
   display: flex;
   gap: 36px;
   justify-content: center;
@@ -90,7 +79,6 @@
 .what-we-offer-text {
   display: flex;
   flex-direction: column;
-  /* align-items: center; */
   text-align: left;
   width: 350px;
   gap: 24px;
@@ -103,104 +91,14 @@
 .interest-form-tabs {
   scroll-margin-top: 60px;
   padding: 12px;
-}
-
-.what-we-offer-new {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  position: relative;
-  width: 100%;
-}
-
-.what-we-offer-grid {
-  display: grid;
-  grid-template: 1fr 1fr / 1fr 1fr 1fr;
-  gap: 24px;
-  justify-items: center;
-}
-
-.what-we-offer-grid-box {
-  width: 300px;
-  aspect-ratio: 5/3;
-  padding: 20px;
-  display: flex;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-  border: 2px solid var(--burnt-orange);
-  background-color: rgba(255, 255, 255, 0.65);
-  border-radius: 12px;
-  font-weight: bolder;
-  cursor: pointer;
-}
-
-.what-we-offer-grid span {
-  display: none;
-  position: absolute;
-  width: 300px;
-  height: 150px;
-  padding: 20px;
-}
-
-.new-thing {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.what-we-offer-grid span.active {
-  display: flex;
-  color: white;
-  align-items: center;
-  justify-content: center;
-}
-
-.what-we-offer-modal {
-  z-index: 5;
-  position: absolute;
-  display: none;
-}
-
-.what-we-offer-modal.active {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.what-we-offer-grid p:hover {
-  background-color: rgba(0, 0, 0, 0.8);
-  color: transparent;
-}
-
-.what-we-offer-grid-mobile {
-  display: none;
-}
-
-@media (hover: none) {
-  .what-we-offer-grid-mobile {
-    display: grid;
-    grid-template: 1fr 1fr / 1fr 1fr 1fr;
-    gap: 24px;
-    justify-items: center;
-  }
-
-  .what-we-offer-grid {
-    display: none;
-  }
-
-  .what-we-offer-grid p:hover {
-    background-color: rgba(255, 255, 255, 0.65);
-    color: var(--black);
-  }
-}
+} */
 
 @media (max-width: 992px) {
   .bookings-container h3 {
     width: 700px;
   }
 
-  .what-we-offer {
+  /* .what-we-offer {
     gap: 24px;
   }
 
@@ -211,7 +109,7 @@
   .what-we-offer-text {
     width: 300px;
     gap: 16px;
-  }
+  } */
 
   .bookings-cards-container {
     gap: 16px;
@@ -232,17 +130,6 @@
 
   .bookings-card p {
     height: 150px;
-  }
-
-  .what-we-offer-grid-box,
-  .what-we-offer-grid span {
-    width: 225px;
-  }
-
-  .what-we-offer-new,
-  .what-we-offer-grid,
-  .what-we-offer-grid-mobile {
-    gap: 20px;
   }
 }
 
@@ -286,7 +173,7 @@
     width: 500px;
   }
 
-  .what-we-offer {
+  /* .what-we-offer {
     gap: 16px;
   }
 
@@ -297,7 +184,7 @@
   .what-we-offer-text {
     width: 250px;
     gap: 16px;
-  }
+  } */
 
   .bookings-cards-container {
     flex-direction: column;
@@ -376,7 +263,7 @@
     width: 375px;
   }
 
-  .what-we-offer {
+  /* .what-we-offer {
     gap: 8px;
     flex-direction: column;
   }
@@ -389,7 +276,7 @@
     width: 400px;
     gap: 8px;
     align-items: center;
-  }
+  } */
 
   .bookings-card {
     grid-template-rows: 40px 160px;
@@ -420,14 +307,14 @@
 }
 
 @media (max-width: 420px) {
-  .what-we-offer img {
+  /* .what-we-offer img {
     width: 250px;
   }
 
   .what-we-offer-text {
     width: 300px;
     gap: 4px;
-  }
+  } */
 
   .bookings-card {
     grid-template-rows: 36px 150px;
